+++ CHANGELOG.md
@@ -2,3 +2,7 @@
 Initial release of Pictograms.
 ## 1.0.1
 Minor bugs fixed.
+## 1.0.2
+Add Automatically FontSize
+Add Size Features
+Add Colors Setting